+++ Samples/iOS-Swift/iOS-SwiftUITests/ProfilingUITests.swift
@@ -15,9 +15,7 @@ final class ProfilingUITests: BaseUITest {
         app.tabBars["Tab Bar"].buttons["Transactions"].tap()
         app.buttons["Start transaction (main thread)"].afterWaitingForExistence("Couldn't find button to start transaction").tap()
 
-        app.tabBars["Tab Bar"].buttons["Extra"].tap()
         app.buttons["ANR filling run loop"].afterWaitingForExistence("Couldn't find button to ANR").tap()
-        app.tabBars["Tab Bar"].buttons["Transactions"].tap()
         app.buttons["Stop transaction"].afterWaitingForExistence("Couldn't find button to end transaction").tap()
 
         let textField = app.textFields["io.sentry.ui-tests.profile-marshaling-text-field"]
